pyproject.toml
[project]
name = "demo-app-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "boto3>=1.42.83",
    "fastapi>=0.135.3",
    "fastapi-storages>=0.4.0",
    "itsdangerous>=2.2.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "sqladmin>=0.24.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.43.0",
]
